Split parsernode util tests to their own respective file
